refactor: extract logic from progress bar to hook
